# docs/openapi.yaml
openapi: 3.0.3

info:
  title: Employee Management API
  version: 1.0.0
  description: API for managing employee records.

servers:
  - url: http://127.0.0.1:8000/api/v1

paths:

  /employees:
    get:
      summary: View all employees
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          example: juan

        - name: department
          in: query
          required: false
          schema:
            type: string
          example: IT

      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'

    post:
      summary: Add a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
            example:
              first_name: Juan
              last_name: Dela Cruz
              email: juan@example.com
              department: IT
              position: Programmer

      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'

        '422':
          description: Validation error

  /employees/{id}:
    get:
      summary: View one employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: Employee found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'

        '404':
          description: Employee not found

    put:
      summary: Update employee information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'

      responses:
        '200':
          description: Employee updated successfully

        '404':
          description: Employee not found

        '422':
          description: Validation error

    delete:
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: Employee deleted successfully

        '404':
          description: Employee not found

components:

  schemas:

    EmployeeInput:
      type: object
      required:
        - first_name
        - last_name
        - email
        - department
        - position

      properties:
        first_name:
          type: string
          example: Juan

        last_name:
          type: string
          example: Dela Cruz

        email:
          type: string
          format: email
          example: juan@example.com

        department:
          type: string
          example: IT

        position:
          type: string
          example: Programmer

    Employee:
      allOf:
        - $ref: '#/components/schemas/EmployeeInput'

        - type: object
          properties:
            id:
              type: integer
              example: 1

            created_at:
              type: string
              format: date-time
              example: 2026-08-04T10:00:00Z

            updated_at:
              type: string
              format: date-time
              example: 2026-08-04T10:00:00Z